eduquiz/requirements.txt
Django==4.2.7
Pillow==10.1.0
python-decouple==3.8
whitenoise==6.6.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0

# Pour l'IA et le traitement des PDF
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0

# Pour les tâches asynchrones (optionnel)
celery==5.3.4
redis==5.0.1

# Pour la production (optionnel)
gunicorn==21.2.0
psycopg2-binary==2.9.9
